+++ differt/src/differt/scene/_triangle_scene.py
@@ -48,6 +48,43 @@ else:
     SionnaScene = Any
 
 
+_B = Bool[Array, " *batch"]
+_F = Float[Array, " *batch"]
+_M = _B | _F
+
+
+@overload
+def _compute_paths(
+    mesh: TriangleMesh,
+    tx_vertices: Float[Array, "num_tx_vertices 3"],
+    rx_vertices: Float[Array, "num_rx_vertices 3"],
+    path_candidates: Int[Array, "num_path_candidates order"],
+    *,
+    epsilon: Float[ArrayLike, " "] | None,
+    hit_tol: Float[ArrayLike, " "] | None,
+    min_len: Float[ArrayLike, " "] | None,
+    smoothing_factor: Float[ArrayLike, " "],
+    confidence_threshold: Float[ArrayLike, " "],
+    batch_size: int | None,
+) -> Paths[_F]: ...
+
+
+@overload
+def _compute_paths(
+    mesh: TriangleMesh,
+    tx_vertices: Float[Array, "num_tx_vertices 3"],
+    rx_vertices: Float[Array, "num_rx_vertices 3"],
+    path_candidates: Int[Array, "num_path_candidates order"],
+    *,
+    epsilon: Float[ArrayLike, " "] | None,
+    hit_tol: Float[ArrayLike, " "] | None,
+    min_len: Float[ArrayLike, " "] | None,
+    smoothing_factor: None,
+    confidence_threshold: Float[ArrayLike, " "],
+    batch_size: int | None,
+) -> Paths[_B]: ...
+
+
 @eqx.filter_jit
 def _compute_paths(
     mesh: TriangleMesh,
@@ -59,11 +96,11 @@ def _compute_paths(
     hit_tol: Float[ArrayLike, ""] | None,
     min_len: Float[ArrayLike, ""] | None,
     smoothing_factor: Float[ArrayLike, ""] | None,
-    confidence_threshold: Float[ArrayLike, ""] = 0.5,
+    confidence_threshold: Float[ArrayLike, ""],
     batch_size: int | None,
-) -> Paths:
+) -> Paths[_M]:
     if min_len is None:
-        dtype = jnp.result_type(mesh.vertices, tx_vertices, tx_vertices)
+        dtype = jnp.result_type(mesh.vertices, tx_vertices, rx_vertices)
         min_len = 10 * jnp.finfo(dtype).eps
 
     min_len = jnp.asarray(min_len)
@@ -254,8 +291,7 @@ def _compute_paths(
     )
 
     if smoothing_factor is not None:
-        mask = None
-        confidence = jnp.stack(
+        mask = jnp.stack(
             (
                 inside_triangles,
                 valid_reflections,
@@ -266,9 +302,8 @@ def _compute_paths(
             axis=-1,
         ).min(axis=-1, initial=1.0)
         if active_rays is not None:
-            confidence *= active_rays
+            mask *= active_rays
     else:
-        confidence = None
         mask = inside_triangles & valid_reflections & ~blocked & ~too_small & is_finite
         if active_rays is not None:
             mask &= active_rays
@@ -306,7 +341,6 @@ def _compute_paths(
         vertices,
         objects,
         mask=mask,
-        confidence=confidence,
         confidence_threshold=confidence_threshold,
     )
 
@@ -323,6 +357,7 @@ def _compute_paths_sbr(
     max_dist: Float[ArrayLike, ""],
     batch_size: int | None,
 ) -> SBRPaths:
+    # TODO: type annotations for SBRPaths with mask dtype
     # 1 - Prepare arrays
 
     # [num_triangles 3 3]
@@ -792,7 +827,26 @@ class TriangleScene(eqx.Module):
         confidence_threshold: Float[ArrayLike, ""] = ...,
         batch_size: int | None = ...,
         disconnect_inactive_triangles: bool = ...,
-    ) -> Paths: ...
+    ) -> Paths[_F]: ...
+
+    @overload
+    def compute_paths(
+        self,
+        order: int | None = None,
+        *,
+        method: Literal["exhaustive"] = "exhaustive",
+        chunk_size: None = None,
+        num_rays: int = int(1e6),
+        path_candidates: Int[ArrayLike, "num_path_candidates order"] | None = None,
+        epsilon: Float[ArrayLike, " "] | None = None,
+        hit_tol: Float[ArrayLike, " "] | None = None,
+        min_len: Float[ArrayLike, " "] | None = None,
+        max_dist: Float[ArrayLike, " "] = 1e-3,
+        smoothing_factor: None = None,
+        confidence_threshold: Float[ArrayLike, " "] = 0.5,
+        batch_size: int | None = 512,
+        disconnect_inactive_triangles: bool = False,
+    ) -> Paths[_B]: ...
 
     @overload
     def compute_paths(
@@ -811,7 +865,26 @@ class TriangleScene(eqx.Module):
         confidence_threshold: Float[ArrayLike, ""] = ...,
         batch_size: int | None = ...,
         disconnect_inactive_triangles: bool = ...,
-    ) -> Paths: ...
+    ) -> Paths[_F]: ...
+
+    @overload
+    def compute_paths(
+        self,
+        order: int,
+        *,
+        method: Literal["hybrid"],
+        chunk_size: None = None,
+        num_rays: int = int(1e6),
+        path_candidates: None = None,
+        epsilon: Float[ArrayLike, " "] | None = None,
+        hit_tol: Float[ArrayLike, " "] | None = None,
+        min_len: Float[ArrayLike, " "] | None = None,
+        max_dist: Float[ArrayLike, " "] = 1e-3,
+        smoothing_factor: None = None,
+        confidence_threshold: Float[ArrayLike, " "] = 0.5,
+        batch_size: int | None = 512,
+        disconnect_inactive_triangles: bool = False,
+    ) -> Paths[_B]: ...
 
     @overload
     def compute_paths(
@@ -830,7 +903,26 @@ class TriangleScene(eqx.Module):
         confidence_threshold: Float[ArrayLike, ""] = ...,
         batch_size: int | None = ...,
         disconnect_inactive_triangles: bool = ...,
-    ) -> SizedIterator[Paths]: ...
+    ) -> SizedIterator[Paths[_F]]: ...
+
+    @overload
+    def compute_paths(
+        self,
+        order: int | None = None,
+        *,
+        method: Literal["exhaustive"] = "exhaustive",
+        chunk_size: int,
+        num_rays: int = int(1e6),
+        path_candidates: None = None,
+        epsilon: Float[ArrayLike, " "] | None = None,
+        hit_tol: Float[ArrayLike, " "] | None = None,
+        min_len: Float[ArrayLike, " "] | None = None,
+        max_dist: Float[ArrayLike, " "] = 1e-3,
+        smoothing_factor: None = None,
+        confidence_threshold: Float[ArrayLike, " "] = 0.5,
+        batch_size: int | None = 512,
+        disconnect_inactive_triangles: bool = False,
+    ) -> SizedIterator[Paths[_B]]: ...
 
     @overload
     def compute_paths(
@@ -849,7 +941,26 @@ class TriangleScene(eqx.Module):
         confidence_threshold: Float[ArrayLike, ""] = ...,
         batch_size: int | None = ...,
         disconnect_inactive_triangles: bool = ...,
-    ) -> Iterator[Paths]: ...
+    ) -> Iterator[Paths[_F]]: ...
+
+    @overload
+    def compute_paths(
+        self,
+        order: int,
+        *,
+        method: Literal["hybrid"],
+        chunk_size: int,
+        num_rays: int = int(1e6),
+        path_candidates: None = None,
+        epsilon: Float[ArrayLike, " "] | None = None,
+        hit_tol: Float[ArrayLike, " "] | None = None,
+        min_len: Float[ArrayLike, " "] | None = None,
+        max_dist: Float[ArrayLike, " "] = 1e-3,
+        smoothing_factor: None = None,
+        confidence_threshold: Float[ArrayLike, " "] = 0.5,
+        batch_size: int | None = 512,
+        disconnect_inactive_triangles: bool = False,
+    ) -> Iterator[Paths[_B]]: ...
 
     @overload
     def compute_paths(
@@ -868,7 +979,26 @@ class TriangleScene(eqx.Module):
         confidence_threshold: Float[ArrayLike, ""] = ...,
         batch_size: int | None = ...,
         disconnect_inactive_triangles: bool = ...,
-    ) -> Paths: ...
+    ) -> Paths[_F]: ...
+
+    @overload
+    def compute_paths(
+        self,
+        order: int | None = None,
+        *,
+        method: Literal["exhaustive"] = "exhaustive",
+        chunk_size: int,
+        num_rays: int = int(1e6),
+        path_candidates: Int[ArrayLike, "num_path_candidates order"],
+        epsilon: Float[ArrayLike, " "] | None = None,
+        hit_tol: Float[ArrayLike, " "] | None = None,
+        min_len: Float[ArrayLike, " "] | None = None,
+        max_dist: Float[ArrayLike, " "] = 1e-3,
+        smoothing_factor: None = None,
+        confidence_threshold: Float[ArrayLike, " "] = 0.5,
+        batch_size: int | None = 512,
+        disconnect_inactive_triangles: bool = False,
+    ) -> Paths[_B]: ...
 
     @overload
     def compute_paths(
@@ -905,7 +1035,7 @@ class TriangleScene(eqx.Module):
         confidence_threshold: Float[ArrayLike, ""] = 0.5,
         batch_size: int | None = 512,
         disconnect_inactive_triangles: bool = False,
-    ) -> Paths | SizedIterator[Paths] | Iterator[Paths] | SBRPaths:
+    ) -> Paths[_M] | SizedIterator[Paths[_M]] | Iterator[Paths[_M]] | SBRPaths:
         """
         Compute paths between all pairs of transmitters and receivers in the scene, that undergo a fixed number of interaction with objects.
 
@@ -1040,6 +1170,7 @@ class TriangleScene(eqx.Module):
 
         .. [#f1] Passing the squared length/distance is useful to avoid computing square root values, which is expensive.
         """
+        # TODO: fix overload matching issue with ty
         if (order is None) == (path_candidates is None):
             msg = "You must specify one of 'order' or `path_candidates`, not both."
             raise ValueError(msg)
@@ -1144,12 +1275,12 @@ class TriangleScene(eqx.Module):
             path_candidates_iter = graph.all_paths_array_chunks(
                 from_=from_,
                 to=to,
-                depth=order + 2,  # type: ignore[reportOptionalOperand]
+                depth=order + 2,  # type: ignore[unsupported-operator]
                 include_from_and_to=False,
                 chunk_size=chunk_size,
             )
             it = (
-                _compute_paths(
+                _compute_paths(  # type: ignore[no-matching-overload]
                     self.mesh,
                     tx_vertices,
                     rx_vertices,
@@ -1176,7 +1307,7 @@ class TriangleScene(eqx.Module):
                 graph.all_paths_array(
                     from_=from_,
                     to=to,
-                    depth=order + 2,  # type: ignore[reportOptionalOperand]
+                    depth=order + 2,  # type: ignore[unsupported-operator]
                     include_from_and_to=False,
                 ),
                 dtype=int,
@@ -1189,7 +1320,7 @@ class TriangleScene(eqx.Module):
             if self.mesh.assume_quads:
                 path_candidates -= path_candidates % 2
 
-        return _compute_paths(
+        return _compute_paths(  # type: ignore[no-matching-overload]
             self.mesh,
             tx_vertices,
             rx_vertices,